requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
gunicorn
python-dotenv
pytest
pytest-django
factory-boy
inertia-django
django-vite